### 2021-08-30 #74 @adolgert
The paper is in `paper/paper.tex`. I don't see whedon commands to debug this. Would it hurt if I resubmitted the paper? I see that the base github link is for a branch, not the actual base.
