### 2021-09-15 #78 @whedon
[REVIEW]: UnitTestDesign.jl: Combinatorial design for unit tests
**Submitting author:** @adolgert (<a href="http://orcid.org/0000-0002-6477-7981">Andrew Dolgert</a>)
**Repository:** <a href="https://github.com/adolgert/UnitTestDesign.jl" target ="_blank">https://github.com/adolgert/UnitTestDesign.jl</a>
**Version:** v0.3.0
