### 2021-09-15 #78 @matbesancon
@odow, @oxinabox thanks for accepting to review this paper. This issue is where the main review happens, in particular on the paper content.
More specific issues can be opened on the package repository directly and linked to this one, in particular when linked to the code
