### 2021-09-15 #78 @matbesancon
The checkboxes above are meant to help getting things started but should not be considered exhaustive
