+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"man/ipog/#In-parameter-order-General","page":"IPOG","title":"In-parameter-order General","text":"","category":"section"},{"location":"man/ipog/#Overview","page":"IPOG","title":"Overview","text":"","category":"section"},{"location":"man/ipog/","page":"IPOG","title":"IPOG","text":"There is an in-parameter-order general (IPOG) algorithm that is published, and I've extended it to be more user-friendly. I'll describe both here.","category":"page"},{"location":"man/ipog/#The-covering-problem","page":"IPOG","title":"The covering problem","text":"","category":"section"},{"location":"man/ipog/","page":"IPOG","title":"IPOG","text":"A combinatorial covering set is a way to select combinations of values for testing. We assume that there are param_cnt items to select. These could be parameter values to send to a function. They could be configuration values for a large simulation. They could be combinations of hardware to test together. We can represent a selected test case as an array of integers, [2, 4, 3, 2], which represents the second selection of the first item, the first selection of the second item, and so on. A test set is a list of test cases.","category":"page"},{"location":"man/ipog/","page":"IPOG","title":"IPOG","text":"We say that a test set has n-way covering when we can choose any n parameters and choose any possible combination of values of those parameters, and find them in some test case of the test set. We can think of the n-way combinations as their own data structure, a covering set, where we might represent it as an array that uses 0 to indicate paramters that aren't in the cover.","category":"page"},{"location":"man/ipog/","page":"IPOG","title":"IPOG","text":"[[1, 1, 0, 0],\n [2, 1, 0, 0],\n [1, 2, 0, 0],\n [0, 1, 1, 0]\n ...\n]","category":"page"},{"location":"man/ipog/","page":"IPOG","title":"IPOG","text":"A test set covers these tuples if every set of non-zero values in the covering set can be found in some test case of the test set.","category":"page"},{"location":"man/ipog/#IPOG","page":"IPOG","title":"IPOG","text":"","category":"section"},{"location":"man/ipog/","page":"IPOG","title":"IPOG","text":"Previous algorithms constructed a single, complete set of parameter values, called a test case, one at a time. This algorithm starts with a few parameters, finds a covering set for them, and then uses this as the basis to add a parameter at a time.","category":"page"},{"location":"man/ipog/","page":"IPOG","title":"IPOG","text":"There are a few sources for this work, which I annotate here.","category":"page"},{"location":"man/ipog/","page":"IPOG","title":"IPOG","text":"Lei, Yu, Raghu Kacker, D. Richard Kuhn, Vadim Okun, and James Lawrence. 2008. “IPOG/IPOG-D: Efficient Test Generation for Multi-Way Combinatorial Testing.” Software Testing, Verification & Reliability 18 (3): 125–48. - This is the most direct description.\nForbes, Michael, Jim Lawrence, Yu Lei, Raghu N. Kacker, and D. Richard Kuhn. 2008. “Refining the In-Parameter-Order Strategy for Constructing Covering Arrays.” Journal of Research of the National Institute of Standards and Technology 113 (5): 287–97. - This updates the algorithm to be more efficient and includes details not explained elsewhere.\nTai, Kuo-Chung, and Yu Lei. 2002. “A Test Generation Strategy for Pairwise Testing.” IEEE Transactions on Software Engineering 28 (1). - The original paper is interesting because it thinks more theoretically about the problem.\nKuhn, D. Richard, Raghu N. Kacker, and Yu Lei. 2013. Introduction to Combinatorial Testing. CRC Press. This book includes an algorithm description, but it may mix up one of the loops.","category":"page"},
[…]
