### 2021-10-19 f32c74d
in the introduction, clarified that I assume you're doing software engineering for testability and that I don't want to talk about that here
