+++ paper/ref.bib
@@ -156,6 +156,16 @@
   location  = "Portland, Oregon"
 }
 
+@inproceedings{chowdhary2009practicing,
+  title={Practicing testability in the real world},
+  author={Chowdhary, Vishal},
+  booktitle={2009 International Conference on Software Testing Verification and Validation},
+  pages={260--268},
+  year={2009},
+  organization={IEEE},
+  doi={10.1109/ICST.2009.53}
+}
+
 @ARTICLE{Segura2016-qh,
   title    = "A Survey on Metamorphic Testing",
   author   = "Segura, S and Fraser, G and Sanchez, A B and Ruiz-Cort{\'e}s, A",
