+++ paper/paper.tex
@@ -280,7 +280,7 @@ There are plenty of improvements to make to these algorithms, such as the use of
 Now that we've explored combinatorial interaction testing, let's return to place it in the context of other similar testing techniques. Compared to other methods for automated test case generation, \cit makes fewer tests, with some wayness guarantees.
 
 \vskip 6pt
-Random testing also generates test cases, and we know, from Arcuri et~al, that these test cases increase code coverage with increasing numbers of test cases in a predictable way~\cite{Arcuri2012-az}. Random testing has the benefit that you can run it for an arbitrarily long time during acceptance testing. There is also no requirement that values be completely random within the domain of arguments. They are almost always biased towards values that carry risk, such as empty strings or other corner cases. Here, a preference for combinatorial interaction testing comes from a lack of test-running resources or from the ease of specifying sample argument values instead of the complete domain of the arguments.
+Random testing also generates test cases, and we know, from Arcuri et~al, that these test cases increase code coverage with increasing numbers of test cases in a predictable way~\cite{Arcuri2012-az}. Random testing has the benefit that you can run it for an arbitrarily long time during acceptance testing. There is also no requirement that values be completely random within the domain of arguments. They are almost always biased towards values that carry risk, such as empty strings or other corner cases. Combinatorial interaction testing is preferable to random testing when there are few test-running resources.
 
 \vskip 6pt
 There are also more mathematical test generation methods, such as orthogonal arrays or Sobol sequences~\cite{He2013-th}. These give an even distribution of values within high-dimensional spaces. They would be excellent to, for instance, seed an optimization problem with local minima, but they aren't designed to explore the space of execution paths of code in the way that \cit is.
