+++ paper/paper.tex
@@ -124,6 +124,9 @@ It's rare that a function expects arguments that really take only a few possible
 An \emph{equivalence class} is a set of input values that would discover the same faults in the code. For instance, if a given function fails for \verb|arg3<2.7|, then it will fail for \verb|arg3=2| as well as \verb|arg3=2.5|, so we consider those equivalent with respect to finding faults. You can choose them by looking at the code or by looking at a specification for the function, but it's possible to miss an equivalence class and miss the faults discovered by that equivalence class.
 
 \vskip 6pt
+A weakness of \cit is that your tests are only as good as the argument values you choose. Even if you choose argument values such that some combination of them would find a failure, it's possible that the subset of test cases generated by the \cit might not include that combination. There is no guarantee that using automated testing to generate hundreds of tests will find a problem.
+
+\vskip 6pt
 One way to defend against our uncertainty about equivalence classes is to combine combinatorial test generation with random testing. Decide that each argument value in the \cit specification represents a generator of values from an equivalence class. Then each test case, output from \cit, is a list of generators of random arguments. This introduces bias, and control, into random test selection. It points to treating test case generation as a continuum of techniques.
 
 \vskip 6pt
