+++ paper/paper.tex
@@ -297,8 +297,11 @@ Property-based testing, for instance, offers a specification to describe the dom
 \vskip 6pt
 Concolic testing performs the remarkable feat of executing a test case and recording that execution in order to learn which argument values control choices at each decision in the code. Then the concolic algorithm chooses a different argument value and tries the test again. This is an automated way to learn corner cases in the code, which should help with the challenge of deciding equivalence classes for arguments. At the same time, concolic testing generally uses logic solvers to decide equivalence classes, and these can be idiosyncratic for mathematical code~\cite{King1976-jt,Wang2018-xh}. The hardest part about concolic testing may be running it, conveniently, within the unit testing framework.
 
+\vskip 6pt
+All of these approaches to automated test generation create more tests with less code than manual alternatives. This makes a test suite less brittle as the code-under-test is modified. It can also create test suites with a reassuring number of test cases, but none of the automated techniques above remove risk. We still need to rely on software engineering techniques, such as tracking of historical bug report rates, in order to understand the maturity of a code base.
+
 \section{Conclusion}
-Combinatorial interaction testing is an advanced technique because it's tailored to a difficult situation, when you can't afford not to spend the time to make your best guess at test cases. However, this kind of automation is easy to specify and run in Julia's ecosystem. The most difficult part of the transition to automated test case generation is shifting from manually-computed checks of test results to checks that work for any combination of arguments, as described in Sec.~\ref{sec:results}. A lot of the strategies in Sec.~\ref{sec:how-to-use} apply to any kind of automated test case generation. They point to the broad applicability of automation for testing and the effectiveness of tools that offer control to the test author.
+Combinatorial interaction testing is an advanced technique because it's tailored to a difficult situation, when you can't afford not to spend the time to make your best guess at test cases. However, this kind of automation is easy to specify and run in Julia's ecosystem. The most difficult part of the transition to automated test case generation is shifting from manually-computed checks of test results to checks that work for any combination of arguments, as described in Sec.~\ref{sec:results}. A lot of the strategies in Sec.~\ref{sec:how-to-use} apply to any kind of automated test case generation. They point to the broad applicability of automation for testing and to the effectiveness of tools that offer control to the test author.
 
 \input{bib.tex}
 
