### 2021-10-20 4d96246
adding the doi for each paper that has one
