+++ paper/ref.bib
@@ -35,7 +35,8 @@
               variables;Algorithm design and analysis;Generators;Coupon
               collector;random testing;theory;Schur
               function;predictability;partition testing;adaptive random
-              testing."
+              testing.",
+  doi      = {10.1109/TSE.2011.121}
 }
 
 
@@ -50,7 +51,8 @@
   month     =  jan,
   year      =  2011,
   keywords  = "Software testing; combinatorial testing (CT); covering array;
-               test case generation"
+               test case generation",
+  doi       = {10.1145/1883612.1883618}
 }
 
 @ARTICLE{Cohen1997-lb,
@@ -62,7 +64,8 @@
   volume  =  23,
   number  =  7,
   pages   = "437--444",
-  year    =  1997
+  year    =  1997,
+  doi     = {10.1109/32.605761}
 }
 
 @ARTICLE{Petke2015-ex,
@@ -107,7 +110,8 @@
               circuits;Combinatorial Interaction
               Testing;Prioritisation;Empirical Studies;Software
               Testing;Combinatorial interaction
-              testing;prioritisation;empirical studies;software testing"
+              testing;prioritisation;empirical studies;software testing",
+  doi       = {10.1109/TSE.2015.2421279}
 }
 
 @ARTICLE{Richard_Kuhn2008-ut,
@@ -139,7 +143,8 @@
   volume    =  100,
   number    =  1,
   pages     = "254--260",
-  year      =  2013
+  year      =  2013,
+  doi       = {10.1093/biomet/ass065}
 }
 
 @INPROCEEDINGS{Cohen2003-pg,
@@ -153,7 +158,8 @@
   series    = "ICSE '03",
   year      =  2003,
   address   = "Washington, DC, USA",
-  location  = "Portland, Oregon"
+  location  = "Portland, Oregon",
+  doi       = {10.1109/ICSE.2003.1201186}
 }
 
 @inproceedings{chowdhary2009practicing,
@@ -195,7 +201,8 @@
   keywords = "program testing;metamorphic testing;test oracle;test
               execution;metamorphic relations;Testing;Search
               engines;Google;Libraries;Concrete;Distance measurement;Web
-              services;Metamorphic testing;oracle problem;survey"
+              services;Metamorphic testing;oracle problem;survey",
+  doi      =  {10.1109/TSE.2016.2532875}
 }
 
 @ARTICLE{Lei2008-xt,
@@ -224,7 +231,8 @@
   number   =  3,
   pages    = "125--148",
   month    =  sep,
-  year     =  2008
+  year     =  2008,
+  doi      =  {10.1002/stvr.381}
 }
 
 @ARTICLE{King1976-jt,
@@ -234,7 +242,8 @@
   volume  =  19,
   number  =  7,
   pages   = "385--394",
-  year    =  1976
+  year    =  1976,
+  doi     =  {10.1145/360248.360252}
 }
 
 @INPROCEEDINGS{Torres-Jimenez2013-kk,
@@ -269,7 +278,8 @@
                methods;metaheuristic methods;algebraic
                methods;Encoding;Generators;Simulated annealing;Markov
                processes;Genetic algorithms;Metals;covering arrays; methods to
-               construct covering arrays"
+               construct covering arrays",
+  doi       =  {10.1109/SYNASC.2013.10}
 }
 
 @INPROCEEDINGS{Inozemtseva2014-gz,
@@ -311,7 +321,8 @@
   year      =  2014,
   address   = "New York, NY, USA",
[…]
