### 2021-11-06 #78 @adolgert
This paper might be ready.
- I made modifications to respond to all issues.
- There is a DOI for every reference that has one.
The reviewer comments were all helpful and positive. Thanks to both of you.
