### 2021-11-08 #78 @oxinabox
>  As the reviewer I confirm that I have read the [JuliaCon conflict of interest policy](https://proceedings.juliacon.org/guide/authors#coi)  and that there are no conflicts of interest for me to review this work.

That URL is broken
