### 2021-11-08 #78 @odow
The links are:
https://proceedings.juliacon.org/about#code_of_conduct
https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy
but perhaps the COI should be https://proceedings.juliacon.org/about#ethics
