### 2021-12-02 #78 @adolgert
Yes, the proof looks good to me. Thanks @matbesancon, @odow and @oxinabox. I appreciate your time.
