### 2021-12-05 #78 @matbesancon
@adolgert can you release a new version (I see there are commits since the last once) and archive it on Zenodo? Once it's done, please post here both the new version and the DOI from Zenodo
see https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance).
