+++ v0.3.0/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"contributing/#Contributing","page":"Contributing","title":"Contributing","text":"","category":"section"},{"location":"contributing/#Examples-of-contributions","page":"Contributing","title":"Examples of contributions","text":"","category":"section"},{"location":"contributing/","page":"Contributing","title":"Contributing","text":"Adding an algorithm for test case generation.\nTelling me which package already has an algorithm to use.\nImproving the documentation or testing builds.\nSuggesting a better API.","category":"page"},{"location":"contributing/","page":"Contributing","title":"Contributing","text":"The Github site has an Issues page, and my email is listed.","category":"page"},{"location":"contributing/#Branch-process","page":"Contributing","title":"Branch process","text":"","category":"section"},{"location":"contributing/","page":"Contributing","title":"Contributing","text":"The trunk is called main. I use PRs from branches or other repos to main. If this project needs more than that, let's discuss it.","category":"page"},{"location":"contributing/#Conduct","page":"Contributing","title":"Conduct","text":"","category":"section"},{"location":"contributing/","page":"Contributing","title":"Contributing","text":"Let's follow the Contributor Covenant Code of Conduct.","category":"page"},{"location":"man/guide/#Guide","page":"Guide","title":"Guide","text":"","category":"section"},{"location":"man/guide/#Kinds-of-test-generation","page":"Guide","title":"Kinds of test generation","text":"","category":"section"},{"location":"man/guide/","page":"Guide","title":"Guide","text":"Combinatorial coverage - Makes the shortest test sets to cover all options.\nall_values: Every value is used at least once.\nall_pairs: Every pair of values is used at least once.\nall_triples: Every triple of values is used at least once.\nall_tuples: Generate test sets with arbitrary coverage level.\nExcursions from a single parameter set\nvalues_excursion: Tries each parameter, one at a time.\npairs_excursion: Tries pairs of parameters, two at a time.\ntriples_excursion: Tries three parameters away from the original set.\nall_tuples: This function works for general excursions, too.\nFull factorial - Every combinations of parameters.\nfull_factorial: Generates all combinations of parameters, filtering those that aren't permitted.","category":"page"},{"location":"man/guide/#Same-interface-for-all-test-generators","page":"Guide","title":"Same interface for all test generators","text":"","category":"section"},{"location":"man/guide/#Increase-coverage-for-subsets-of-parameters","page":"Guide","title":"Increase coverage for subsets of parameters","text":"","category":"section"},{"location":"man/guide/","page":"Guide","title":"Guide","text":"Sometimes there are a few parameters that are more important to test fully. In that case, choose a base-level coverage for the whole test set, such as the 2-way coverage of all-pairs. Then pass a separate argument to request that the subset of parameters have greater coverage.","category":"page"},{"location":"man/guide/","page":"Guide","title":"Guide","text":"For instance, this requrests that the first, third, and fourth parameters have 3-way coverage, meaning full-factorial, while the second parameter has only 2-way coverage. This is more meaningful when there are lots of parameters.","category":"page"},{"location":"man/guide/","page":"Guide","title":"Guide","text":"using UnitTestDesign  # hide\ntest_set = all_pairs(\n    [1, 2, 3], [\"low\", \"mid\" ,\"high\"], [1.0, 3.7, 4.9], [:greedy, :relax, :optim];\n    wayness = Dict(3 => [[1, 3, 4]])\n    )","category":"page"},{"location":"man/guide/","page":"Guide","title":"Guide","text":"The wayness argument is a dictionary, where the dictionary keys are the higher coverage levels, so \"3\" means all triples. The dictionary values are
[…]
