### 2021-12-05 #78 @adolgert
Got it. The DOI is 10.5281/zenodo.579923. The new version is v0.3.0. (I didn't go for 1.0 because there is an upcoming paper, and I'll make upgrades to the algorithm.)
* Release: https://github.com/adolgert/UnitTestDesign.jl/releases
* Zenodo: https://zenodo.org/record/5759923#.Ya0ZmXuIYRU
I should have enabled zenodo on the repo and then done the release. I went the other way around, which was less automatic, but it will be automatic next time.
