### 2021-12-05 #78 @matbesancon
@adolgert can you check the DOIs indicated as missing in your references?
