### 2021-12-05 #78 @adolgert
Darn it. Those weren't on the papers, but they are on the owners' pages. Sorry.
