### 2021-12-05 #78 @adolgert
FWIW, the donations link is broken. It is now https://numfocus.org/donate-to-joss, and I made a donation b/c this has been great.
